+++ Assets/yohkan/runtime/scripts/YohkanAssetProvider_Extender.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: db0c2e06e2384b699fd952969b7d8632
+timeCreated: 1731135153
