+++ Assets/yohkan/runtime/scripts/interfaces/IAssetExtender.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: e7d6aed305d346ca92f8db2d6838eb82
+timeCreated: 1731134535
