+++ Assets/yohkan/runtime/scripts/internals/AssetInfo.cs
@@ -8,6 +8,7 @@ namespace yohkan.runtime.scripts.internals
         public string Address;
         public AssetReference Reference;
         public UnityEngine.Object Asset;
+        public UnityEngine.Sprite SpriteAsset;//　Texture2d -> Sprite変換が必要になるのでSpriteだけ例外的にフィールドを分ける
         public Type AssetType;
     }
 }
